# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/v2/auth/identity/users/login": {
            "post": {
                "description": "LoginByEmail",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Identity"
                ],
                "summary": "LoginByEmail",
                "parameters": [
                    {
                        "description": "LoginByUsernameRequest",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user-service_api_dto.LoginByUsernameRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/user-service_api_helper.BaseHttpResponse"
                        }
                    },
                    "400": {
                        "description": "Failed",
                        "schema": {
                            "$ref": "#/definitions/user-service_api_helper.BaseHttpResponse"
                        }
                    },
                    "409": {
                        "description": "Failed",
                        "schema": {
                            "$ref": "#/definitions/user-service_api_helper.BaseHttpResponse"
                        }
                    }
                }
            }
        },
        "/v2/auth/identity/users/register": {
            "post": {
                "description": "RegisterByEmail",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Identity"
                ],
                "summary": "RegisterByEmail",
                "parameters": [
                    {
                        "description": "RegisterUserByUsernameRequest",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user-service_api_dto.RegisterUserByUsernameRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/user-service_api_helper.BaseHttpResponse"
                        }
                    },
                    "400": {
                        "description": "Failed",
                        "schema": {
                            "$ref": "#/definitions/user-service_api_helper.BaseHttpResponse"
                        }
                    },
                    "409": {
                        "description": "Failed",
                        "schema": {
                            "$ref": "#/definitions/user-service_api_helper.BaseHttpResponse"
                        }
                    }
                }
            }
        },
...
        "/v2/auth/resource/phones/": {
            "put": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Update a Phone",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Phones"
                ],
                "summary": "Update a Phone",
                "parameters": [
                    {
                        "description": "Update a Phone",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user-service_api_dto.UpdatePhoneRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Phone response",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/user-service_api_helper.BaseHttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/user-service_api_dto.PhoneResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/user-service_api_helper.BaseHttpResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Create a Phone",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Phones"
                ],
                "summary": "Create a Phone",
                "parameters": [
                    {
                        "description": "Create a Phone",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user-service_api_dto.CreatePhoneRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Phone response",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/user-service_api_helper.BaseHttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/user-service_api_dto.PhoneResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/user-service_api_helper.BaseHttpResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "user-service_api_dto.CreatePhoneRequest": {
            "type": "object",
            "required": [
                "mobileNumber"
            ],
            "properties": {
                "country": {
                    "type": "string",
                    "maxLength": 11,
                    "minLength": 3
                },
                "mobileNumber": {
                    "type": "string",
                    "maxLength": 11,
                    "minLength": 11
                }
            }
        },
...
        "user-service_api_dto.LoginByUsernameRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "minLength": 6
                },
                "password": {
                    "type": "string",
                    "minLength": 6
                }
            }
        },
        "user-service_api_dto.PhoneResponse": {
            "type": "object",
            "properties": {
                "mobileNumber": {
                    "type": "string"
                }
            }
        },
...
        "user-service_api_dto.RegisterUserByUsernameRequest": {
            "type": "object",
            "required": [
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "minLength": 6
                },
                "password": {
                    "type": "string",
                    "minLength": 6
                }
            }
        },
...
        "user-service_api_dto.UpdatePhoneRequest": {
            "type": "object",
            "required": [
                "mobileNumber"
            ],
            "properties": {
                "country": {
                    "type": "string",
                    "maxLength": 11,
                    "minLength": 3
                },
                "mobileNumber": {
                    "type": "string",
                    "maxLength": 11,
                    "minLength": 11
                }
            }
        },
...
        "user-service_api_helper.BaseHttpResponse": {
            "type": "object",
            "properties": {
                "error": {},
                "result": {},
                "resultCode": {
                    "$ref": "#/definitions/user-service_api_helper.ResultCode"
                },
                "success": {
                    "type": "boolean"
                },
                "validationErrors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/user-service_api_validation.ValidationError"
                    }
                }
            }
        },
        "user-service_api_helper.ResultCode": {
            "type": "integer",
            "enum": [
                0,
                40001,
                40101,
                40301,
                40401,
                42901,
                42902,
                50001,
                50002
            ],
            "x-enum-varnames": [
                "Success",
                "ValidationError",
                "AuthError",
                "ForbiddenError",
                "NotFoundError",
                "LimiterError",
                "OtpLimiterError",
                "CustomRecovery",
                "InternalError"
            ]
        },
        "user-service_api_validation.ValidationError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "property": {
                    "type": "string"
                },
                "tag": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "AuthBearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}